requirements.txt
streamlit
pydub
numpy
scipy
matplotlib
plotly
captcha
pillow
pytest